Split php.spec.ts in two
